% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{da21uid_statcan_v2021}
\alias{da21uid_statcan_v2021}
\title{StatCan dissemination area ID lookup}
\format{
A data table with the columns:
\describe{
\item{da21uid}{Dissemination area IDs in Canada}
}
}
\usage{
data(da21uid_statcan_v2021)
}
\description{
StatCan dissemination area ID lookup
}
\keyword{datasets}
