% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_erintervention.R
\name{dummy_erintervention_mri}
\alias{dummy_erintervention_mri}
\title{Generate simulated erintervention data}
\usage{
dummy_erintervention_mri(
  dbcon = NULL,
  nid = 1000,
  n_hospitals = 10,
  int_code = NULL,
  cohort = NULL,
  seed = NULL
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to a GEMINI database, required to look up intervention codes.
Required when \code{int_code} is missing.}

\item{nid}{(\code{integer})\cr Number of unique encounter IDs to simulate.
Encounter IDs may repeat, resulting in a data table with more rows than \code{nid}.
Optional when \code{cohort} is provided.}

\item{n_hospitals}{(\code{integer})\cr Number of hospitals to simulate.
Optional when \code{cohort} is provided.}

\item{int_code}{(\verb{character or vector})\cr Optional, user-specified intervention codes to include in the returned
data table. Required when \code{dbcon} is not provided.}

\item{cohort}{(\verb{data.frame or data.table})\cr Optional, data frame or data table containing the fields:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID number
\item \code{hospital_num} (\code{integer}): Mock hospital ID number
When \code{cohort} is not NULL, \code{nid} and \code{n_hospitals} are ignored.
}}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results}
}
\value{
(\code{data.table})\cr A data.table object similar to the "ipintervention" table that contains the columns:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID number; integers starting from 1 or from \code{cohort}
\item \code{hospital_num} (\code{integer}): Mock hospital ID number; integers starting from 1 or from \code{cohort}
\item \code{intervention_code} (\code{character}): A valid CCI code(s) describing the services (procedures/intervention)
performed for or on behalf of the patient to improve health. For this simulation, it will be for an MRI.
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "erintervention" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.

This function simulates data with CCI codes detailing the type of intervention used in the emergency department.
}
\examples{
dummy_erintervention_mri(nid = 1000, int_code = c("3AN40VA", "3SC40WC"))
dummy_erintervention_mri(cohort = dummy_ipadmdad(), int_code = "3AN40VA")

}
