% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_ipintervention_mri_maid}
\alias{dummy_ipintervention_mri_maid}
\title{Generate simulated ipintervention data}
\usage{
dummy_ipintervention_mri_maid(
  dbcon = NULL,
  n = 1000,
  n_hospitals = 10,
  cohort = NULL,
  int_code = NULL,
  seed = NULL
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to a GEMINI database, required to look up intervention codes.
Required when \code{int_code} is missing, otherwise not used if \code{int_code} is provided.}

\item{cohort}{(\verb{data.frame or data.table})\cr Optional, data frame or data table containing the columns:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID number
\item \code{hospital_num} (\code{integer}): Mock hospital ID number
When \code{cohort} is not NULL, \code{nid} and \code{n_hospitals} are ignored.
}}

\item{int_code}{(\code{character})\cr Optional, user-specified intervention codes to include in the returned data table.
Required when \code{dbcon} is missing.}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results}

\item{nid}{(\code{integer})\cr Number of unique encounter IDs to simulate.
Encounter IDs may repeat, resulting in a data table with more rows than \code{nid}.
It is not used if \code{cohort} is provided.}
}
\value{
(\code{data.table})\cr A data.table object similar to the "ipintervention" table that contains the columns:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID number; integers starting from 1 or provided from \code{cohort}
\item \code{hospital_id} (\code{integer}): Mock hospital ID number; integers starting from 1 or provided from \code{cohort}
\item \code{intervention_code} (\code{character}): A valid CCI code(s) describing the services (procedures/intervention)
performed for or on behalf of the patient to improve health.
For this simulation, it will either be for an MRI or medical assistance in dying (MAID)
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "ipintervention" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.

This function simulates data with CCI codes detailing the type of intervention that occurred in an inpatient stay.
}
\examples{
dummy_ipintervention_mri_maid(nid = 1000, int_code = c("3AN40VA", "3SC40WC"))
dummy_ipintervention_mri_maid(nid = 1000, int_code = "3SC40WC")

}
