% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_lab.R
\name{dummy_lab_cbc_electrolyte}
\alias{dummy_lab_cbc_electrolyte}
\title{Generate simulated lab data}
\usage{
dummy_lab_cbc_electrolyte(
  nid = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  cohort = NULL,
  seed = NULL
)
}
\arguments{
\item{nid}{(\code{integer}) Number of unique encounter IDs to simulate. In this data table, each ID occurs once.
It is ignored if \code{cohort} is provided.}

\item{n_hospitals}{(\code{integer}) Number of hospitals in simulated dataset.
It is ignored if \code{cohort} is provided}

\item{time_period}{(\code{numeric}): Date range of data, by years or specific dates in either format:
("yyyy-mm-dd", "yyyy-mm-dd") or (yyyy, yyyy).
It is ignored if \code{cohort} is provided.}

\item{cohort}{(\verb{data.frame or data.table})\cr Optional, a data frame or data table with columns:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID numbers
\item \code{hospital_num} (\code{integer}): Mock hospital ID numbers
\item \code{admission_date_time} (\code{character}): Date and time of IP admission in YYYY-MM-DD HH:MM format
\item \code{discharge_date_time} (\code{character}): Date and time of IP discharge in YYYY-MM-DD HH:MM format.
When \code{cohort} is not NULL, \code{nid}, \code{n_hospitals}, and \code{time_period} are ignored.
}}

\item{seed}{(\code{integer}) Optional, a number for setting the seed to get reproducible results.}
}
\value{
(\code{data.table})\cr A data.table object similar to the "lab" table that contains the following fields:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID; integers starting from 1 or as seen in \code{cohort}
\item \code{hospital_num} (\code{integer}): Mock hospital ID; integers starting from 1 or as seen in \code{cohort}
\item \code{test_type_mapped_omop} (\code{character}):	Test name and code mapped by GEMINI; currently two tests are available: 3000963 (hemoglobin) and 3019550 (sodium)
\item \code{test_name_raw} (\code{character}): Test name as reported by hospital
\item \code{test_code_raw} (\code{character}): Test code as reported by hospital
\item \code{result_value} (\code{character}): Test results
\item \code{collection_date_time} (\code{character}):	Date and time when the sample was collected
}
}
\description{
This function creates a synthetic dataset with a subset of lab tests that
are contained in the GEMINI "lab" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
The function currently focuses on simulating two lab tests: hemoglobin and sodium, as they are often used to identify routine blood work tests of complete blood count and electrolytes.
This function will return: collection date time, information about the test type, test code, and test result value.
It is a long format data table.
}
\examples{
dummy_lab_cbc_electrolyte(nid = 10, n_hospitals = 1, seed = 1)
dummy_lab_cbc_electrolyte(cohort = dummy_ipadmdad())

}
