% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_cci}
\alias{lookup_cci}
\title{CCI lookup table for MAID and MRI codes}
\description{
This is the section of the CCI intervention lookup table with codes for MRI and MAID.
Its columns are :
\itemize{
\item \code{intervention_code} (\code{character}): CCI intervention code
\item \code{maid} (\code{logical}): TRUE if the code is a MAID code, FALSE otherwise
}
}
