% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_cci_mri-data.R
\docType{data}
\name{lookup_cci_mri}
\alias{lookup_cci_mri}
\title{Lookup table for MRI CCI codes}
\format{
A data frame with X rows and Y columns:
\describe{
\item{code}{CCI code}
\item{description}{Description}
...
}
}
\source{
Generated from internal GEMINI datasets.
}
\usage{
lookup_cci_mri
}
\description{
Contains the mapping of MRI procedure codes to CCI categories.
}
\keyword{datasets}
