% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_cci_mri_maid.R
\docType{data}
\name{lookup_cci_mri_maid}
\alias{lookup_cci_mri_maid}
\title{MRI CCI lookup table - MAID variant}
\format{
A data table with columns:
\describe{
\item{intervention_code}{CCI intervention code}
\item{cci_short_title}{Short title}
\item{cci_long_title}{Long title}
}
}
\usage{
lookup_cci_mri_maid
}
\description{
MRI CCI lookup table - MAID variant
}
\keyword{datasets}
