% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rjohnson_trunc}
\alias{rjohnson_trunc}
\title{Internal function to sample from the t distribution truncated within a given range}
\usage{
rjohnson_trunc(n, min, max)
}
\arguments{
\item{n}{(\code{integer})\cr The length of the final vector}

\item{min}{(\code{numeric})\cr The minimum for truncating the distribution}

\item{max}{(\code{numeric})\cr The maximum for truncating the distribution}
}
\value{
A numeric vector following the specified Johnson distribution.
}
\description{
This function samples a numeric vector from the Johnson distribution as per the params and returns it.
This is used to sample CBC lab test result values.
}
\keyword{internal}
