% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_cci}
\alias{lookup_cci}
\title{CCI lookup table for MAID and MRI codes}
\format{
A data table with the columns:
\describe{
\item{intervention_code}{CCI intervention code}
\item{maid}{TRUE if the code is a MAID code, FALSE otherwise}
}
}
\usage{
data(lookup_cci)
}
\description{
CCI lookup table for MAID and MRI codes
}
\keyword{datasets}
