% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{simulate_data_tables}
\alias{simulate_data_tables}
\title{Wrapper function to call other data simulation functions.}
\usage{
simulate_data_tables(
  tables,
  nid = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  ...
)
}
\arguments{
\item{tables}{(\code{vector})\cr A \code{character} vector listing the names of required data tables}

\item{nid}{(\code{integer})\cr The number of mock encounter IDs to simulate.}

\item{n_hospitals}{(\code{integer})\cr The number of mock hospital ID numbers to simulate.}

\item{time_period}{(\code{vector})\cr A numeric or character vector containing the data range of the data
by years or specific dates in either format: ("yyyy-mm-dd", "yyyy-mm-dd") or (yyyy, yyyy).
The start date and end date will be (yyyy-01-01 and yyyy-12-31) if (yyyy, yyyy)
is the date range format provided.}

\item{...}{Additional arguments that may be passed into data simulation functions.}
}
\value{
(\code{list})\cr A list containing \code{data.table} objects,
one per entry in the \code{tables} parameter where entries are valid GEMINI data tables.
}
\description{
This function works in conjunction with all other data simulation functions.
The user is required to input a list of data tables they want to create,
in addition to other inputs to customize data tables.
It returns a list of data tables, considering inter-table relations, creating a full database.
}
