% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{simulate_data_tables}
\alias{simulate_data_tables}
\title{Wrapper function that calls data simulation functions to create a synthetic,customizable database that reflects the inter-table relations of the GEMINI database.}
\usage{
simulate_data_tables(
  tables,
  nid = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  ...
)
}
\arguments{
\item{tables}{(\code{vector})\cr A \code{character} vector listing the names of required data tables}

\item{nid}{(\code{integer})\cr The number of mock encounter IDs to simulate.}

\item{n_hospitals}{(\code{integer})\cr The number of mock hospital ID numbers to simulate.}

\item{time_period}{(\code{vector})\cr A numeric or character vector containing the data range of the data
by years or specific dates in either format: ("yyyy-mm-dd", "yyyy-mm-dd") or (yyyy, yyyy).
The start date and end date will be (yyyy-01-01 and yyyy-12-31) if (yyyy, yyyy)
is the date range format provided.}

\item{...}{Additional arguments that may be passed into data simulation functions.
These arguments are normally used to customize table outputs.}
}
\value{
(\code{list})\cr A list containing \code{data.table} objects,
one per entry in the \code{tables} parameter where entries are valid GEMINI data tables.
May include: \code{admdad}, \code{ipscu}, \code{er}, \code{erdiagnosis}, \code{ipdiagnosis}, \code{locality_variables},
\code{lab}, \code{radiology}, \code{erintervention}, \code{ipintervention}, \code{transfusion}, \code{physicians}
}
\description{
A wrapper that coordinates table-specific simulation functions to generate
relational synthetic tables that reflect the inter-table structure of the GEMINI data.
Users specify which tables to generate and provide shared inputs such as the
number of encounters, hospitals, and the time period.
The function returns a list of simulated \code{data.table}s with inter-table
relationships handled automatically. Specifically, the \code{admdad} table is generated first
and provides the encounter IDs used as the primary key for subsequent table.
All tables are simulated to mirror their real-world linkage patterns to the
\code{admdad} table in GEMINI data.
Available tables include:
\itemize{
\item \code{admdad}
\item \code{ipscu}
\item \code{er}
\item \code{erdiagnosis}
\item \code{ipdiagnosis}
\item \code{locality_variables}
\item \code{lab} *: currently simulates CBC or electrolyte tests only
\item \code{radiology} *: currently simulates MRI, CT, and ultrasound imaging data only
\item \code{erintervention} *: currently simulates intervention MRI
\item \code{ipintervention} *: currently simulates interventions MRI and MAID only
\item \code{transfusion}: transfusion information about blood product and issue date times
\item \code{physicians}
}

See \href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary} for
table definitions and individual simulation function documentation for details.
}
